added swal to confirm delete is really wanted

+++ src/components/ArticleHub.js
@@ -4,6 +4,7 @@ import styled from "styled-components";
 import { withState } from "../MyState";
 import ReactPaginate from "react-paginate";
 import axios from "axios";
+import sweetie from "sweetalert2";
 
 const HubWrapper = styled.div`
   width: 100vw;
@@ -27,20 +28,34 @@ class ArticleHub extends Component {
     const toKeep = this.state.articlesToPaginate.filter(article => {
       return article._id !== id;
     });
-    axios.delete(`/api/articles/${id}`).then(res => {
-      this.setState(
-        {
-          articlesToPaginate: toKeep
-        },
-        () => {
-          this.getSomeArticles();
+    sweetie
+      .fire({
+        title: "Are you sure?! Are you sure?!",
+        text:
+          "Are you sure?! Are you sure?! Are you sure?! Are you sure?! Are you sure?! Are you sure?! Are you sure?!",
+        showCancelButton: true,
+        confirmButtonColor: "#610707",
+        cancelButtonColor: "rgba(109,108,108,0.9)",
+        cancelButtonText: "NO!",
+        confirmButtonText: "Delete!",
+        padding: "2.5rem"
+      })
+      .then(result => {
+        if (result.value) {
+          axios.delete(`/api/articles/${id}`).then(res => {
+            this.setState(
+              {
+                articlesToPaginate: toKeep
+              },
+              () => {
+                this.getSomeArticles();
+              }
+            );
+          });
         }
-      );
-    });
+      });
   };
-
   getSomeArticles = page => {
-    window.scrollTo(0, 0);
     axios
       .get(`/api/articles/some/${page}`)
       .then(res => {
